@@ -24,6 +24,8 @@ import android.widget.Spinner;
 import android.widget.TextView;
 import android.Manifest;
 
+import com.filipe.footballmatch.Repositories.EventRepository;
+import com.filipe.footballmatch.Repositories.UserRepository;
 import com.filipe.footballmatch.Utilities.DatePickerFragment;
 import com.filipe.footballmatch.Models.Event;
 import com.filipe.footballmatch.Utilities.MessageDialog;
@@ -35,6 +37,7 @@ import com.google.android.gms.common.ConnectionResult;
 import com.google.android.gms.common.GooglePlayServicesNotAvailableException;
 import com.google.android.gms.common.GooglePlayServicesRepairableException;
 import com.google.android.gms.common.api.GoogleApiClient;
+import com.google.android.gms.location.FusedLocationProviderClient;
 import com.google.android.gms.location.LocationListener;
 import com.google.android.gms.location.LocationRequest;
 import com.google.android.gms.location.LocationServices;
@@ -43,6 +46,7 @@ import com.google.android.gms.location.places.Places;
 import com.google.android.gms.location.places.ui.PlacePicker;
 import com.google.android.gms.maps.model.LatLng;
 import com.google.android.gms.maps.model.LatLngBounds;
+import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.firebase.database.DataSnapshot;
 import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
@@ -54,11 +58,14 @@ import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.Date;
 
+import butterknife.BindView;
+import butterknife.ButterKnife;
+
 import static android.R.attr.value;
 import static com.filipe.footballmatch.R.id.buttonCancel;
 
 public class CreateMatchActivity extends AppCompatActivity implements
-        GoogleApiClient.OnConnectionFailedListener, LocationListener, GoogleApiClient.ConnectionCallbacks {
+        GoogleApiClient.OnConnectionFailedListener, GoogleApiClient.ConnectionCallbacks, UserRepository.OnGetUsersList, EventRepository.OnFinished {
 
     private static final String TAG = "CreateMatchActivity";
     private static final int PLACE_PICKER_REQUEST = 1;
@@ -66,23 +73,65 @@ public class CreateMatchActivity extends AppCompatActivity implements
     private static final int PERMISSION_REQUEST_CODE = 100;
     private static final int ADD_PLAYER_REQUEST_CODE = 1002;
 
+    UserRepository userRepository = new UserRepository();
+    EventRepository eventRepository = new EventRepository();
+
     private GoogleApiClient mGoogleApiClient;
     private LocationRequest mLocationRequest;
 
-    private LinearLayout section2;
-    private RelativeLayout section3;
-    private LinearLayout playerListLayout;
-    private TextView mName;
-    private TextView mAddress;
-    private TextView mPhone;
-    private Spinner mSpinner;
+    @BindView(R.id.section2)
+    LinearLayout section2;
+
+    @BindView(R.id.section3)
+    RelativeLayout section3;
+
+    @BindView(R.id.event_players_list)
+    LinearLayout playerListLayout;
+
+    @BindView(R.id.venue_name)
+    TextView mName;
+
+    @BindView(R.id.venue_address)
+    TextView mAddress;
+
+    @BindView(R.id.venue_phone)
+    TextView mPhone;
+
+    @BindView(R.id.spinner)
+    Spinner mSpinner;
+
+    @BindView(R.id.pickerButton)
     TextView pickerButton;
 
-    private TextView mTime;
-    private TextView mDate;
+    @BindView(R.id.event_time)
+    TextView mTime;
+
+    @BindView(R.id.event_date)
+    TextView mDate;
+
+    @BindView(R.id.date_picker)
+    TextView datePicker;
+
+    @BindView(R.id.time_picker)
+    TextView timePicker;
+
+    @BindView(R.id.tilEventName)
+    TextInputLayout eventNameLayout;
+
+    @BindView(R.id.callButton)
+    TextView callButton;
+
+    @BindView(R.id.buttonAddPlayer)
+    TextView addPlayerButton;
+
+    @BindView(R.id.buttonCreateEvent)
+    TextView createEventButton;
+
+    @BindView(R.id.buttonCancel)
+    TextView cancelButton;
 
     Place place;
-    TextInputLayout eventNameLayout;
+
     private String eventName;
     private String placeName;
     private Date eventDate;
@@ -93,139 +142,80 @@ public class CreateMatchActivity extends AppCompatActivity implements
     private String id;
     private boolean pickerClicked = false;
 
+    private FusedLocationProviderClient mFusedLocationClient;
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
 
         // The content layout of screen is set
         setContentView(R.layout.activity_create_match);
+        ButterKnife.bind(this);
 
         // The action bar title is customized
         getSupportActionBar().setDisplayOptions(ActionBar.DISPLAY_SHOW_CUSTOM);
         getSupportActionBar().setCustomView(R.layout.action_bar);
 
-        // The layout is now built
-        // First, the Layouts that will act as sections for the screen are set
-        section2 = (LinearLayout) findViewById(R.id.section2);
-        section3 = (RelativeLayout) findViewById(R.id.section3);
-        playerListLayout = (LinearLayout) findViewById(R.id.event_players_list);
+        mFusedLocationClient = LocationServices.getFusedLocationProviderClient(this);
 
-        // Then, the TextInputLayout field that the user to give a name for the event
-        eventNameLayout = (TextInputLayout) findViewById(R.id.tilEventName);
-
-        // The TextViews that will display the info about the venue
-        mName = (TextView) findViewById(R.id.venue_name);
-        mAddress = (TextView) findViewById(R.id.venue_address);
-        mPhone = (TextView) findViewById(R.id.venue_phone);
-
-        // The pickers that will allow the user to choose the date...
-        TextView datePicker = (TextView) findViewById(R.id.date_picker);
-        mDate = (TextView) findViewById(R.id.event_date);
-
-        // ... and the time of the match
-        TextView timePicker = (TextView) findViewById(R.id.time_picker);
-        mTime = (TextView) findViewById(R.id.event_time);
-
-        // The spinner that will allow the user to select the number of players in the match
-        mSpinner = (Spinner) findViewById(R.id.spinner);
+        // The spinner that will allow the user to select the number of players in the match is set
         String[] items = new String[]{"10 (5x2)", "14 (7x2)", "22 (11x2)"};
         ArrayAdapter<String> adapter = new ArrayAdapter<>(this, R.layout.item_spinner, items);
         mSpinner.setAdapter(adapter);
 
-        // And, finally, the TextViews that will act as LoginActivity screen buttons are set
-        TextView callButton = (TextView) findViewById(R.id.callButton);
-        pickerButton = (TextView) findViewById(R.id.pickerButton);
-        TextView addPlayerButton = (TextView) findViewById(R.id.buttonAddPlayer);
-        TextView createEventButton = (TextView) findViewById(R.id.buttonCreateEvent);
-        TextView cancelButton = (TextView) findViewById(buttonCancel);
-
-        // The actual user id is recovered from the device local storage...
-        SharedPreferences saved_values = PreferenceManager.getDefaultSharedPreferences(getApplicationContext());
-        id = saved_values.getString(getString(R.string.user_id_SharedPref), "");
+        // The actual user id is recovered
+        id = userRepository.getUidCurrentUser();
 
         // Click Listener for button choose match venue
-        pickerButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                pickerClicked = true;
-                getLocation();
-            }
+        pickerButton.setOnClickListener(v -> {
+            pickerClicked = true;
+            getLocation();
         });
 
         // Click Listener for button call venue
-        callButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Utility.call(CreateMatchActivity.this, place.getPhoneNumber().toString());
-            }
-        });
+        callButton.setOnClickListener(v -> Utility.call(CreateMatchActivity.this, place.getPhoneNumber().toString()));
 
         // Click Listener for timePicker, opens time picker fragment
-        timePicker.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                FragmentManager fm = getFragmentManager();
-                DialogFragment newFragment = new TimePickerFragment();
-                newFragment.show(fm, "timePicker");
-            }
+        timePicker.setOnClickListener(v -> {
+            FragmentManager fm = getFragmentManager();
+            DialogFragment newFragment = new TimePickerFragment();
+            newFragment.show(fm, "timePicker");
         });
 
         // Click Listener for datePicker, opens date picker fragment
-        datePicker.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                FragmentManager fm = getFragmentManager();
-                DialogFragment newFragment = new DatePickerFragment();
-                newFragment.show(fm, "datePicker");
-            }
+        datePicker.setOnClickListener(v -> {
+            FragmentManager fm = getFragmentManager();
+            DialogFragment newFragment = new DatePickerFragment();
+            newFragment.show(fm, "datePicker");
         });
 
         // Click Listener for button add player
-        addPlayerButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent intent = new Intent(CreateMatchActivity.this, ListUsersActivity.class);
-                intent.putExtra("isFromCreateMatch", true);
-                CreateMatchActivity.this.startActivityForResult(intent, ADD_PLAYER_REQUEST_CODE);
-            }
+        addPlayerButton.setOnClickListener(v -> {
+            Intent intent = new Intent(CreateMatchActivity.this, ListUsersActivity.class);
+            intent.putExtra("isFromCreateMatch", true);
+            CreateMatchActivity.this.startActivityForResult(intent, ADD_PLAYER_REQUEST_CODE);
         });
 
         // Click Listener for button create event
-        createEventButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                if (validateData()) {
-                    if (Utility.isConnectedToNet(CreateMatchActivity.this)) {
-                        registerMatch();
-                    } else {
-                        Utility.noNetworkError(CreateMatchActivity.this);
-                    }
+        createEventButton.setOnClickListener(v -> {
+            if (validateData()) {
+                if (Utility.isConnectedToNet(CreateMatchActivity.this)) {
+                    registerMatch();
                 } else {
-                    displayErrorPopup();
+                    Utility.noNetworkError(CreateMatchActivity.this);
                 }
+            } else {
+                displayErrorPopup();
             }
         });
 
         // Click Listener for button cancel
-        cancelButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                final MessageDialog dialog = new MessageDialog(CreateMatchActivity.this, R.string.cancel_register_match_message, -1, R.string.dialog_edit_no_text, R.string.dialog_edit_yes_text);
-                dialog.setCancelable(false);
-                dialog.show();
-                dialog.noButton.setOnClickListener(new View.OnClickListener() {
-                    @Override
-                    public void onClick(View view) {
-                        dialog.dismiss();
-                    }
-                });
-                dialog.yesButton.setOnClickListener(new View.OnClickListener() {
-                    @Override
-                    public void onClick(View view) {
-                        finish();
-                    }
-                });
-            }
+        cancelButton.setOnClickListener(v -> {
+            final MessageDialog dialog = new MessageDialog(CreateMatchActivity.this, R.string.cancel_register_match_message, -1, R.string.dialog_edit_no_text, R.string.dialog_edit_yes_text);
+            dialog.setCancelable(false);
+            dialog.show();
+            dialog.noButton.setOnClickListener(view -> dialog.dismiss());
+            dialog.yesButton.setOnClickListener(view -> finish());
         });
     }
 
@@ -271,23 +261,33 @@ public class CreateMatchActivity extends AppCompatActivity implements
                 && resultCode == Activity.RESULT_OK) {
             String userId = data.getStringExtra("userId");
 
+            switch (mSpinner.getSelectedItemPosition()) {
+                case 0:
+                    maxPlayers = 10;
+                    break;
+                case 1:
+                    maxPlayers = 14;
+                    break;
+                case 2:
+                    maxPlayers = 22;
+                    break;
+            }
+
             if (userId == null || userId.isEmpty()) {
                 Utility.generalError(CreateMatchActivity.this, getString(R.string.error_general));
-            }
-            else if (playerIdList != null) {
+            } else if (playerIdList != null) {
                 if (playerIdList.contains(userId)) {
                     Utility.generalError(CreateMatchActivity.this, getString(R.string.error_user_already_added));
                 } else if (playerIdList.size() >= maxPlayers) {
                     Utility.generalError(CreateMatchActivity.this, getString(R.string.error_player_max_number_exceeded));
                 } else {
                     playerIdList.add(userId);
-                    getIdInfo(userId);
+                    userRepository.fetchListUsers(this);
                 }
-            }
-            else {
+            } else {
                 playerIdList = new ArrayList<>();
                 playerIdList.add(userId);
-                getIdInfo(userId);
+                userRepository.fetchListUsers(this);
             }
 
         } else {
@@ -344,8 +344,31 @@ public class CreateMatchActivity extends AppCompatActivity implements
                 ActivityCompat.requestPermissions(CreateMatchActivity.this,
                         new String[]{Manifest.permission.ACCESS_FINE_LOCATION, Manifest.permission.ACCESS_COARSE_LOCATION},
                         PERMISSION_REQUEST_CODE);
+            } else {
+                // LocationServices.FusedLocationApi.requestLocationUpdates(mGoogleApiClient, mLocationRequest, this);
+                mFusedLocationClient.getLastLocation()
+                        .addOnSuccessListener(this, location -> {
+                            // Got last known location. In some rare situations this can be null.
+                            if (location != null) {
+                                LatLngBounds LIKELY_PLACE = new LatLngBounds(new LatLng(location.getLatitude() - 0.003, location.getLongitude() - 0.003)
+                                        , new LatLng(location.getLatitude() + 0.003, location.getLongitude() + 0.003));
+
+                                try {
+
+                                    PlacePicker.IntentBuilder intentBuilder =
+                                            new PlacePicker.IntentBuilder();
+                                    intentBuilder.setLatLngBounds(LIKELY_PLACE);
+                                    Intent intent = intentBuilder.build(CreateMatchActivity.this);
+                                    startActivityForResult(intent, PLACE_PICKER_REQUEST);
+
+                                } catch (GooglePlayServicesRepairableException
+                                        | GooglePlayServicesNotAvailableException e) {
+                                    e.printStackTrace();
+                                    Utility.generalError(CreateMatchActivity.this, e.getMessage());
+                                }
+                            }
+                        });
             }
-            LocationServices.FusedLocationApi.requestLocationUpdates(mGoogleApiClient, mLocationRequest, this);
         }
     }
 
@@ -361,34 +384,6 @@ public class CreateMatchActivity extends AppCompatActivity implements
     @Override
     public void onConnectionSuspended(int i) {
 
-    }
-
-    // When google places API returns a location, the map will show the user surroundings when opened,
-    // allowing him to choose a place to create an event
-    @Override
-    public void onLocationChanged(Location location) {
-        if (location != null) {
-
-            LatLngBounds LIKELY_PLACE = new LatLngBounds(new LatLng(location.getLatitude() - 0.003, location.getLongitude() - 0.003)
-                    , new LatLng(location.getLatitude() + 0.003, location.getLongitude() + 0.003));
-
-            LocationServices.FusedLocationApi.removeLocationUpdates(mGoogleApiClient, CreateMatchActivity.this);
-
-            try {
-
-                PlacePicker.IntentBuilder intentBuilder =
-                        new PlacePicker.IntentBuilder();
-                intentBuilder.setLatLngBounds(LIKELY_PLACE);
-                Intent intent = intentBuilder.build(CreateMatchActivity.this);
-                startActivityForResult(intent, PLACE_PICKER_REQUEST);
-
-            } catch (GooglePlayServicesRepairableException
-                    | GooglePlayServicesNotAvailableException e) {
-                e.printStackTrace();
-                Utility.generalError(CreateMatchActivity.this, e.getMessage());
-            }
-
-        }
     }
 
     // The create event data is validated before submitted to the database
@@ -466,14 +461,6 @@ public class CreateMatchActivity extends AppCompatActivity implements
     // If the data is ok, the match is registered
     public void registerMatch() {
 
-        //  Write a message to the database
-        FirebaseDatabase database = FirebaseDatabase.getInstance();
-        DatabaseReference myRef = database.getReference("Event");
-
-        // Creating new event node, which returns the unique key value
-        // new event node would be /Event/$eventid/
-        String eventId = myRef.push().getKey();
-
         // Creating Event object
         Event event = new Event();
 
@@ -487,68 +474,13 @@ public class CreateMatchActivity extends AppCompatActivity implements
         event.setNumberOfPlayers(mSpinner.getSelectedItem().toString());
         event.setDate(eventDate);
         event.setPlayersIdList(playerIdList);
-        event.setEventKey(eventId);
         event.setCreator(id);
 
-        myRef.child(eventId).setValue(event);
-
-        final MessageDialog dialog = new MessageDialog(CreateMatchActivity.this, R.string.success_register_match, R.string.dialog_edit_ok_text, -1, -1);
-        dialog.setCancelable(false);
-        dialog.show();
-        dialog.okButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                Intent intent = new Intent(CreateMatchActivity.this, MainMenuActivity.class);
-                CreateMatchActivity.this.startActivity(intent);
-
-                dialog.cancel();
-            }
-        });
+        eventRepository.saveEvent(event, this);
     }
 
-    // For every player added to the event, the app search for its info in the database...
-    public void getIdInfo(final String id) {
-
-        FirebaseDatabase database = FirebaseDatabase.getInstance();
-        DatabaseReference myRef = database.getReference("Person/");
-
-        // Read from the database
-        myRef.addListenerForSingleValueEvent(new ValueEventListener() {
-
-            @Override
-            public void onDataChange(DataSnapshot dataSnapshot) {
-
-                for (DataSnapshot dsp : dataSnapshot.getChildren()) {
-                    if (dsp.getKey() != null
-                            && dsp.getKey().equals(id)) {
-                        Person temp = dsp.getValue(Person.class);
-                        temp.setUserKey(dsp.getKey());
-                        playerList.add(temp);
-                        populatePlayerLayout();
-                        Log.d(TAG, "Value is: " + value);
-                    }
-                    else if (dsp.getValue(Person.class).getOldKey() != null
-                            && dsp.getValue(Person.class).getOldKey().equals(id)) {
-                        Person temp = dsp.getValue(Person.class);
-                        temp.setUserKey(dsp.getKey());
-                        playerList.add(temp);
-                        populatePlayerLayout();
-                        Log.d(TAG, "Value is: " + value);
-                    }
-                }
-            }
-
-            @Override
-            public void onCancelled(DatabaseError error) {
-                // Failed to read value
-                Log.w(TAG, "Failed to read value.", error.toException());
-                Utility.generalError(CreateMatchActivity.this, error.getMessage());
-            }
-
-        });
-    }
-
-    // ... and populate the layout with the player name and preferred position
+    // For every player added to the event, the app search for its info in the database
+    // and populate the layout with the player name and preferred position
     public void populatePlayerLayout() {
         View.inflate(this, R.layout.item_user_list, playerListLayout);
 
@@ -561,16 +493,56 @@ public class CreateMatchActivity extends AppCompatActivity implements
         playerPreferredPosition.setText(" (" + playerList.get(i).getPreferredPosition() + ")");
 
 
-        view.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent intent = new Intent(CreateMatchActivity.this, ViewProfileActivity.class);
-                intent.putExtra("userKey", v.getTag().toString());
-                CreateMatchActivity.this.startActivity(intent);
-            }
+        view.setOnClickListener(v -> {
+            Intent intent = new Intent(CreateMatchActivity.this, ViewProfileActivity.class);
+            intent.putExtra("userKey", v.getTag().toString());
+            CreateMatchActivity.this.startActivity(intent);
         });
 
     }
 
 
+    @Override
+    public void OnGetUsersListSuccess(DataSnapshot dataSnapshot) {
+
+        for (DataSnapshot dsp : dataSnapshot.getChildren()) {
+            if ((dsp.getKey() != null && dsp.getKey().equals(id))
+            || (dsp.getValue(Person.class).getOldKey() != null
+                    && dsp.getValue(Person.class).getOldKey().equals(id))){
+
+                Person temp = dsp.getValue(Person.class);
+                temp.setUserKey(dsp.getKey());
+                playerList.add(temp);
+                populatePlayerLayout();
+                Log.d(TAG, "Value is: " + value);
+            }
+        }
+    }
+
+    @Override
+    public void OnGetUsersListFailed(String error) {
+        // Failed to read value
+        Log.e(TAG, error);
+        Utility.generalError(CreateMatchActivity.this, error);
+    }
+
+    @Override
+    public void onEventSaveSuccess() {
+        final MessageDialog dialog = new MessageDialog(CreateMatchActivity.this, R.string.success_register_match, R.string.dialog_edit_ok_text, -1, -1);
+        dialog.setCancelable(false);
+        dialog.show();
+        dialog.okButton.setOnClickListener(view -> {
+            Intent intent = new Intent(CreateMatchActivity.this, MainMenuActivity.class);
+            CreateMatchActivity.this.startActivity(intent);
+
+            dialog.cancel();
+        });
+    }
+
+    @Override
+    public void onEventSaveFailed(String exception) {
+        // Failed to read value
+        Log.e(TAG, exception);
+        Utility.generalError(CreateMatchActivity.this, exception);
+    }
 }
