@@ -1,9 +1,11 @@
 package com.filipe.footballmatch.Activities;
 
+import android.app.Activity;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
+import android.net.Uri;
 import android.os.Bundle;
 import android.preference.PreferenceManager;
 import android.provider.MediaStore;
@@ -18,54 +20,66 @@ import android.widget.ImageView;
 import android.widget.Spinner;
 import android.widget.TextView;
 
+import com.bumptech.glide.Glide;
+import com.filipe.footballmatch.Repositories.UserRepository;
+import com.filipe.footballmatch.Utilities.CustomPhotoPickerDialog;
 import com.filipe.footballmatch.Utilities.MessageDialog;
 import com.filipe.footballmatch.Models.Person;
 import com.filipe.footballmatch.R;
-import com.filipe.footballmatch.Utilities.Utility;
-import com.google.android.gms.tasks.OnFailureListener;
 import com.google.android.gms.tasks.OnSuccessListener;
-import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.database.DatabaseReference;
-import com.google.firebase.database.FirebaseDatabase;
 import com.google.firebase.storage.FileDownloadTask;
 import com.google.firebase.storage.FirebaseStorage;
 import com.google.firebase.storage.StorageReference;
-import com.google.firebase.storage.UploadTask;
 
 import org.parceler.Parcels;
 
-import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.io.IOException;
-import java.util.regex.Pattern;
 
-import static android.R.attr.password;
-import static com.filipe.footballmatch.R.id.tilEmail;
+import butterknife.BindView;
+import butterknife.ButterKnife;
 
 /**
  * Created by Filipe on 21/01/2017.
  */
 
-public class EditProfileActivity extends AppCompatActivity {
+public class EditProfileActivity extends AppCompatActivity implements UserRepository.OnFinished, UserRepository.OnGetUserImage {
 
-    private ImageView profilePicture;
-    private TextInputLayout tilName;
-    private TextInputLayout tilAge;
-    private Spinner spPreferredPosition;
-    private TextInputLayout tilContactNumber;
+    @BindView(R.id.profile_picture_iv)
+    ImageView profilePicture;
 
-    StorageReference profileImageRef;
+    @BindView(R.id.tilName)
+    TextInputLayout tilName;
 
+    @BindView(R.id.tilAge)
+    TextInputLayout tilAge;
+
+    @BindView(R.id.spPreferredPosition)
+    Spinner spPreferredPosition;
+
+    @BindView(R.id.tilContactNumber)
+    TextInputLayout tilContactNumber;
+
+    @BindView(R.id.tilEmail)
+    TextInputLayout tilEmail;
+
+    @BindView(R.id.buttonConfirm)
+    TextView buttonConfirm;
+
+    @BindView(R.id.buttonCancel)
+    TextView buttonCancel;
+
+    private static final int SELECT_PICTURE = 1593;
     final static int REQUEST_IMAGE_CAPTURE = 1001;
 
     public static final String TAG = EditProfileActivity.class.getSimpleName();
+    UserRepository userRepository = new UserRepository();
 
-    File localFile = null;
     Person person;
-    String id;
 
     Person editedInfo = new Person();
 
+    private CustomPhotoPickerDialog photoDialog;
     boolean wasPictureChanged = false;
 
     @Override
@@ -74,6 +88,7 @@ public class EditProfileActivity extends AppCompatActivity {
 
         // The content layout of screen is set
         setContentView(R.layout.activity_edit_profile);
+        ButterKnife.bind(this);
 
         // The action bar title is customized
         getSupportActionBar().setDisplayOptions(ActionBar.DISPLAY_SHOW_CUSTOM);
@@ -82,22 +97,8 @@ public class EditProfileActivity extends AppCompatActivity {
         // Get the person data from the previous activity
         person = Parcels.unwrap(getIntent().getExtras().getParcelable("person"));
 
-        // The layout is now built
-        // First, the TextInputLayouts, the ImageView and the Spinner, that
-        // will allow the user to update the profile info are set
-        profilePicture = (ImageView) findViewById(R.id.profile_picture_iv);
-        tilName = (TextInputLayout) findViewById(R.id.tilName);
-        tilAge = (TextInputLayout) findViewById(R.id.tilAge);
-        spPreferredPosition = (Spinner) findViewById(R.id.spPreferredPosition);
-        tilContactNumber = (TextInputLayout) findViewById(R.id.tilContactNumber);
-
         // the email field will not be used in the edit profile flow
-        TextInputLayout tilEmail = (TextInputLayout) findViewById(R.id.tilEmail);
         tilEmail.setVisibility(View.GONE);
-
-        // Then, the TextViews that will act as LoginActivity screen buttons
-        TextView buttonConfirm = (TextView) findViewById(R.id.buttonConfirm);
-        TextView buttonCancel = (TextView) findViewById(R.id.buttonCancel);
 
         // If the person data recovered from the intent is not null, then the fields are populated
         if (person != null) {
@@ -110,166 +111,79 @@ public class EditProfileActivity extends AppCompatActivity {
             tilAge.getEditText().setText(Integer.toString(person.getAge()));
         }
 
-        // The actual user id is recovered from the device local storage...
-        SharedPreferences saved_values = PreferenceManager.getDefaultSharedPreferences(getApplicationContext());
-        id = saved_values.getString(getString(R.string.user_id_SharedPref), "");
+        // Now, the app recovers the player image from the cloud storage and the player info from the database
+        userRepository.fetchUserImage(userRepository.getUidCurrentUser(), this);
 
-        // Now, the app recovers the player image from the cloud storage
+        // Click Listener for button update picture, the user can get an image from its gallery or take a new one using the camera
+        profilePicture.setOnClickListener( view-> showOptionsPhoto());
 
-        // Get instance of Storage
-        FirebaseStorage storage = FirebaseStorage.getInstance();
+        // Click Listener for button confirm, where the data is validated and saved
+        buttonConfirm.setOnClickListener(view-> {
 
-        // Create a storage reference from our app
-        StorageReference storageRef = storage.getReferenceFromUrl("gs://football-match-a60fb.appspot.com");
+                if (validateInfo()) {
+                    Bitmap bitmap = null;
 
-        if (!id.isEmpty()) {
-            // Create a reference to 'images/id.jpg'
-            profileImageRef = storageRef.child("images/" + id + ".jpg");
+                    if (wasPictureChanged) {
+                        profilePicture.setDrawingCacheEnabled(true);
+                        profilePicture.buildDrawingCache();
+                        bitmap = profilePicture.getDrawingCache();
+                    }
 
-            // Try create a file with the recovered picture
-            try {
-                localFile = File.createTempFile(id, "jpg");
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-        }
-
-        if (localFile != null) {
-            profileImageRef.getFile(localFile).addOnSuccessListener(new OnSuccessListener<FileDownloadTask.TaskSnapshot>() {
-                @Override
-                public void onSuccess(FileDownloadTask.TaskSnapshot taskSnapshot) {
-                    Bitmap myImg = BitmapFactory.decodeFile(localFile.getPath());
-                    profilePicture.setImageBitmap(myImg);
+                    userRepository.saveUser(editedInfo, bitmap, EditProfileActivity.this);
                 }
-            }).addOnFailureListener(new OnFailureListener() {
-                @Override
-                public void onFailure(@NonNull Exception exception) {
-                    // Handle any errors
-//                    Utility.generalError(EditProfileActivity.this, exception.getMessage());
-                }
-            });
-        }
+        });
 
-        // Click Listener for button update picture
-        profilePicture.setOnClickListener(new View.OnClickListener() {
+        // Click Listener for button cancel
+        buttonCancel.setOnClickListener(view -> {
+
+                final MessageDialog dialog = new MessageDialog(EditProfileActivity.this, R.string.cancel_profile_update_message, -1, R.string.dialog_edit_no_text, R.string.dialog_edit_yes_text);
+                dialog.setCancelable(false);
+                dialog.show();
+                dialog.noButton.setOnClickListener(view1 -> dialog.dismiss());
+                dialog.yesButton.setOnClickListener(view12 -> {
+                    Intent intent = new Intent(EditProfileActivity.this, ViewProfileActivity.class);
+                    EditProfileActivity.this.startActivity(intent);
+                    finish();
+                });
+        });
+
+    }
+
+    private void showOptionsPhoto() {
+        photoDialog = new CustomPhotoPickerDialog(this, new CustomPhotoPickerDialog
+                .OnOptionPhotoSelected() {
             @Override
-            public void onClick(View v) {
+            public void onGallery() {
+                Intent intent = new Intent();
+                intent.setType("image/*");
+                intent.setAction(Intent.ACTION_GET_CONTENT);
+                startActivityForResult(Intent.createChooser(intent, "Select Picture"), SELECT_PICTURE);
+                photoDialog.dismiss();
+            }
+
+            @Override
+            public void onCamera() {
+                // Here, thisActivity is the current activity
                 Intent takePictureIntent = new Intent(MediaStore.ACTION_IMAGE_CAPTURE);
                 if (takePictureIntent.resolveActivity(EditProfileActivity.this.getPackageManager()) != null) {
                     startActivityForResult(takePictureIntent, REQUEST_IMAGE_CAPTURE);
                 }
+                photoDialog.dismiss();
             }
         });
-
-        // Click Listener for button confirm
-        buttonConfirm.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-
-                if (validateInfo()) {
-                    if (wasPictureChanged)
-                        addPictureToStorage();
-                    else
-                        updateProfileData();
-                }
-
-            }
-        });
-
-        // Click Listener for button cancel
-        buttonCancel.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                final MessageDialog dialog = new MessageDialog(EditProfileActivity.this, R.string.cancel_profile_update_message, -1, R.string.dialog_edit_no_text, R.string.dialog_edit_yes_text);
-                dialog.setCancelable(false);
-                dialog.show();
-                dialog.noButton.setOnClickListener(new View.OnClickListener() {
-                    @Override
-                    public void onClick(View view) {
-                        dialog.dismiss();
-                    }
-                });
-                dialog.yesButton.setOnClickListener(new View.OnClickListener() {
-                    @Override
-                    public void onClick(View view) {
-                        Intent intent = new Intent(EditProfileActivity.this, ViewProfileActivity.class);
-                        EditProfileActivity.this.startActivity(intent);
-                        finish();
-                    }
-                });
-            }
-        });
-
-    }
-
-    // If the picture was updated, the old one is updated by the new one in the storage
-    public void addPictureToStorage() {
-        // Get the data from an ImageView as bytes
-        profilePicture.setDrawingCacheEnabled(true);
-        profilePicture.buildDrawingCache();
-        Bitmap bitmap = profilePicture.getDrawingCache();
-        ByteArrayOutputStream baos = new ByteArrayOutputStream();
-        bitmap.compress(Bitmap.CompressFormat.JPEG, 100, baos);
-        byte[] data = baos.toByteArray();
-
-        UploadTask uploadTask = profileImageRef.putBytes(data);
-        uploadTask.addOnFailureListener(new OnFailureListener() {
-            @Override
-            public void onFailure(@NonNull Exception exception) {
-                // Handle unsuccessful uploads
-                Log.d(TAG, "error: " + exception.getMessage());
-                final MessageDialog dialog = new MessageDialog(EditProfileActivity.this, exception.getMessage(), R.string.dialog_edit_ok_text, -1, -1);
-                dialog.setCancelable(false);
-                dialog.show();
-                dialog.okButton.setOnClickListener(new View.OnClickListener() {
-                    @Override
-                    public void onClick(View view) {
-                        dialog.cancel();
-                    }
-                });
-            }
-        }).addOnSuccessListener(new OnSuccessListener<UploadTask.TaskSnapshot>() {
-            @Override
-            public void onSuccess(UploadTask.TaskSnapshot taskSnapshot) {
-                // taskSnapshot.getMetadata() contains file metadata such as size, content-type, and download URL.
-//                Uri downloadUrl = taskSnapshot.getDownloadUrl();
-                Log.d(TAG, "success");
-                updateProfileData();
-            }
-        });
-    }
-
-    // If the profile data was changed, the info in the database is updated
-    public void updateProfileData() {
-
-
-
-
-        FirebaseDatabase database = FirebaseDatabase.getInstance();
-        DatabaseReference myRef = database.getReference("Person");
-
-        myRef.child(id).setValue(editedInfo);
-
-        final MessageDialog dialog = new MessageDialog(this, R.string.success_update_profile, R.string.dialog_edit_ok_text, -1, -1);
-        dialog.setCancelable(false);
-        dialog.show();
-        dialog.okButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                dialog.cancel();
-                Intent intent = new Intent(EditProfileActivity.this, ViewProfileActivity.class);
-                EditProfileActivity.this.startActivity(intent);
-                finish();
-            }
-        });
+        photoDialog.show();
     }
 
     @Override
     public void onActivityResult(int requestCode, int resultCode, Intent data) {
-        if (requestCode == REQUEST_IMAGE_CAPTURE && resultCode == EditProfileActivity.this.RESULT_OK) {
+        if (requestCode == REQUEST_IMAGE_CAPTURE && resultCode == RESULT_OK) {
             Bundle extras = data.getExtras();
             Bitmap imageBitmap = (Bitmap) extras.get("data");
             profilePicture.setImageBitmap(imageBitmap);
+            wasPictureChanged = true;
+        } else if (requestCode == SELECT_PICTURE && resultCode == RESULT_OK) {
+            Uri imageUri = data.getData();
+            Glide.with(EditProfileActivity.this).load(imageUri).asBitmap().into(profilePicture);
             wasPictureChanged = true;
         }
     }
@@ -295,8 +209,7 @@ public class EditProfileActivity extends AppCompatActivity {
         if (editedInfo.getName() == null || editedInfo.getName().isEmpty()) {
             tilName.setError(getString(R.string.error_invalid_name));
             validated = false;
-        }
-        else {
+        } else {
             tilName.setErrorEnabled(false);
         }
 
@@ -322,8 +235,7 @@ public class EditProfileActivity extends AppCompatActivity {
 
         if (person.getPreferredPosition() == null) {
             position = 0;
-        }
-        else {
+        } else {
             switch (person.getPreferredPosition()) {
                 case "Goalkeeper":
                     position = 0;
@@ -358,4 +270,37 @@ public class EditProfileActivity extends AppCompatActivity {
         spPreferredPosition.setSelection(position);
     }
 
+    @Override
+    public void onUserSaveSuccess() {
+//        if (mProgressDialog != null) {
+//            mProgressDialog.dismiss();
+//        }
+
+        final MessageDialog dialog = new MessageDialog(this, R.string.success_update_profile, R.string.dialog_edit_ok_text, -1, -1);
+        dialog.setCancelable(false);
+        dialog.show();
+        dialog.okButton.setOnClickListener(view -> {
+            dialog.cancel();
+            Intent intent = new Intent(EditProfileActivity.this, ViewProfileActivity.class);
+            EditProfileActivity.this.startActivity(intent);
+            finish();
+        });
+    }
+
+    @Override
+    public void onUserSaveFailed(String exception) {
+//        if (mProgressDialog != null) {
+//            mProgressDialog.dismiss();
+//        }
+
+        final MessageDialog dialog = new MessageDialog(EditProfileActivity.this, exception, R.string.dialog_edit_ok_text, -1, -1);
+        dialog.setCancelable(false);
+        dialog.show();
+        dialog.okButton.setOnClickListener(view -> dialog.cancel());
+    }
+
+    @Override
+    public void OnGetUserImageSuccess(Bitmap myImg) {
+        profilePicture.setImageBitmap(myImg);
+    }
 }
