@@ -20,6 +20,8 @@ import android.widget.TextView;
 import com.filipe.footballmatch.Adapters.AvailableEventsAdapter;
 import com.filipe.footballmatch.Models.Event;
 import com.filipe.footballmatch.R;
+import com.filipe.footballmatch.Repositories.EventRepository;
+import com.filipe.footballmatch.Repositories.UserRepository;
 import com.filipe.footballmatch.Utilities.DatePickerFragment;
 import com.filipe.footballmatch.Utilities.Utility;
 import com.google.firebase.database.DataSnapshot;
@@ -31,28 +33,49 @@ import com.google.firebase.database.ValueEventListener;
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 
-import static android.R.attr.id;
-import static com.google.android.gms.analytics.internal.zzy.v;
+import butterknife.BindView;
+import butterknife.ButterKnife;
+
 
 /**
  * Created by Filipe on 23/01/2017.
  */
 
-public class ListAvailableEventsActivity extends AppCompatActivity {
+public class ListAvailableEventsActivity extends AppCompatActivity implements EventRepository.OnGetEventsList{
 
-    RecyclerView mRecyclerView;
     AvailableEventsAdapter adapter;
 
+    @BindView(R.id.available_event_recycler_view)
+    RecyclerView mRecyclerView;
+
+    @BindView(R.id.search_event_layout)
     CardView searchEventLayout;
+
+    @BindView(R.id.search_event_showFilter)
     LinearLayout showFilterLayout;
 
+    @BindView(R.id.tilEventName)
     TextInputLayout eventNameLayout;
+
+    @BindView(R.id.tilVenueName)
     TextInputLayout venueNameLayout;
-    private TextView mDate;
+
+    @BindView(R.id.date_picker)
+    TextView datePicker;
+
+    @BindView(R.id.event_date)
+    TextView mDate;
+
+    @BindView(R.id.created_by_checkBox)
     CheckBox created_by_checkBox;
+
+    @BindView(R.id.joined_checkBox)
     CheckBox joined_checkBox;
 
+    @BindView(R.id.search_button)
     TextView searchEventButton;
+
+    @BindView(R.id.show_filter_button)
     TextView showFilterButton;
 
     String eventName;
@@ -61,6 +84,8 @@ public class ListAvailableEventsActivity extends AppCompatActivity {
     String id;
 
     public static final String TAG = ListAvailableEventsActivity.class.getSimpleName();
+    UserRepository userRepository = new UserRepository();
+    EventRepository eventRepository = new EventRepository();
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -68,133 +93,86 @@ public class ListAvailableEventsActivity extends AppCompatActivity {
 
         // The content layout of screen is set
         setContentView(R.layout.activity_list_available_events);
+        ButterKnife.bind(this);
 
         // The action bar title is customized
         getSupportActionBar().setDisplayOptions(ActionBar.DISPLAY_SHOW_CUSTOM);
         getSupportActionBar().setCustomView(R.layout.action_bar);
 
-        // The id of the user is recovered from the Shared Preferences
-        SharedPreferences saved_values = PreferenceManager.getDefaultSharedPreferences(getApplicationContext());
-        id = saved_values.getString(getString(R.string.user_id_SharedPref), "");
-
-        // The layout is now built
-        // First, the CardViews that will act as sections for the screen are set
-        searchEventLayout = (CardView) findViewById(R.id.search_event_layout);
-        showFilterLayout = (LinearLayout) findViewById(R.id.search_event_showFilter);
-
-        // Then, the TextInputLayout fields that will collect the user inputs to filter the list
-        eventNameLayout = (TextInputLayout) findViewById(R.id.tilEventName);
-        venueNameLayout = (TextInputLayout) findViewById(R.id.tilVenueName);
-
-        // The pickers that will allow the user to choose the date...
-        TextView datePicker = (TextView) findViewById(R.id.date_picker);
-        mDate = (TextView) findViewById(R.id.event_date);
-
-        // The checkboxes that will allow the user to filter only matches created or joined by him
-        created_by_checkBox = (CheckBox) findViewById(R.id.created_by_checkBox);
-        joined_checkBox = (CheckBox) findViewById(R.id.joined_checkBox);
-
-        // The TextViews that will act as ListUsersActivity screen buttons are set
-        searchEventButton = (TextView) findViewById(R.id.search_button);
-        showFilterButton = (TextView) findViewById(R.id.show_filter_button);
-
-        // The RecyclerView, that will display the list of elements, is set
-        mRecyclerView = (RecyclerView) findViewById(R.id.available_event_recycler_view);
+        // The id of the user is recovered
+        id = userRepository.getUidCurrentUser();
 
         // when the activity starts, all available events are shown;
-        searchDatabase();
+        eventRepository.fetchListEvents(this);
 
         //Click Listener for button search event
-        searchEventButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
+        searchEventButton.setOnClickListener(v -> {
 
-                // The filter section is hided, a show filter button is set and the list is displayed with the results
-                searchEventLayout.setVisibility(View.GONE);
-                showFilterLayout.setVisibility(View.VISIBLE);
-                mRecyclerView.setVisibility(View.VISIBLE);
+            // The filter section is hided, a show filter button is set and the list is displayed with the results
+            searchEventLayout.setVisibility(View.GONE);
+            showFilterLayout.setVisibility(View.VISIBLE);
+            mRecyclerView.setVisibility(View.VISIBLE);
 
-                eventName = eventNameLayout.getEditText().getText().toString().trim();
-                venueName = venueNameLayout.getEditText().getText().toString().trim();
+            eventName = eventNameLayout.getEditText().getText().toString().trim();
+            venueName = venueNameLayout.getEditText().getText().toString().trim();
 
-                searchDatabase();
-            }
+            eventRepository.fetchListEvents(this);
         });
 
         //Click Listener for button show filter
-        showFilterButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                searchEventLayout.setVisibility(View.VISIBLE);
-                showFilterLayout.setVisibility(View.GONE);
-                mRecyclerView.setVisibility(View.GONE);
-            }
+        showFilterButton.setOnClickListener(v -> {
+            searchEventLayout.setVisibility(View.VISIBLE);
+            showFilterLayout.setVisibility(View.GONE);
+            mRecyclerView.setVisibility(View.GONE);
         });
 
         // Click Listener for datePicker, opens date picker fragment
-        datePicker.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                FragmentManager fm = getFragmentManager();
-                DialogFragment newFragment = new DatePickerFragment();
-                newFragment.show(fm, "datePicker");
-            }
+        datePicker.setOnClickListener(v -> {
+            FragmentManager fm = getFragmentManager();
+            DialogFragment newFragment = new DatePickerFragment();
+            newFragment.show(fm, "datePicker");
         });
 
     }
 
-    public void searchDatabase() {
+    @Override
+    public void OnGetEventsListSuccess(DataSnapshot dataSnapshot) {
+        ArrayList<Event> events = new ArrayList<>();
+        SimpleDateFormat sdf = new SimpleDateFormat("dd-MM-yyyy");
 
-        // An instance of FirebaseDatabase is set
-        FirebaseDatabase database = FirebaseDatabase.getInstance();
-        DatabaseReference myRef = database.getReference("Event/");
+        // Recover all the items from the database that match with the filter
+        for (DataSnapshot dsp : dataSnapshot.getChildren()) {
 
-        // Read from the database
-        myRef.addListenerForSingleValueEvent(new ValueEventListener() {
-
-            @Override
-            public void onDataChange(DataSnapshot dataSnapshot) {
-
-                ArrayList<Event> events = new ArrayList<>();
-                SimpleDateFormat sdf = new SimpleDateFormat("dd-MM-yyyy");
-
-                // Recover all the items from the database that match with the filter
-                for (DataSnapshot dsp : dataSnapshot.getChildren()) {
-
-                    Event temp = dsp.getValue(Event.class);
-                    if ((eventNameLayout.getEditText().getText().toString().trim().isEmpty()
-                            || temp.getEventName().equals(eventName))
-                            && (venueNameLayout.getEditText().getText().toString().trim().isEmpty()
-                            || temp.getName().contains(venueName))
-                            && (mDate.getText().toString().trim().isEmpty()
-                            || sdf.format(temp.getDate()).trim().equals(mDate.getText().toString().trim()))
-                            && (!created_by_checkBox.isChecked() || temp.getCreator().equals(id))
-                            && (!joined_checkBox.isChecked()
-                            || (temp.getPlayersIdList() != null && temp.getPlayersIdList().contains(id)))) {
-                        events.add(temp);
-                    }
-                }
-
-                // set the RecyclerView parameters
-                LinearLayoutManager llm = new LinearLayoutManager(ListAvailableEventsActivity.this);
-                llm.setOrientation(LinearLayoutManager.VERTICAL);
-
-                mRecyclerView.setLayoutManager(llm);
-                mRecyclerView.setHasFixedSize(true);
-
-                // Update the list with the results
-                adapter = new AvailableEventsAdapter(ListAvailableEventsActivity.this, events);
-                mRecyclerView.setAdapter(adapter);
-
+            Event temp = dsp.getValue(Event.class);
+            if ((eventNameLayout.getEditText().getText().toString().trim().isEmpty()
+                    || temp.getEventName().equals(eventName))
+                    && (venueNameLayout.getEditText().getText().toString().trim().isEmpty()
+                    || temp.getName().contains(venueName))
+                    && (mDate.getText().toString().trim().isEmpty()
+                    || sdf.format(temp.getDate()).trim().equals(mDate.getText().toString().trim()))
+                    && (!created_by_checkBox.isChecked() || temp.getCreator().equals(id))
+                    && (!joined_checkBox.isChecked()
+                    || (temp.getPlayersIdList() != null && temp.getPlayersIdList().contains(id)))) {
+                events.add(temp);
             }
+        }
 
-            @Override
-            public void onCancelled(DatabaseError error) {
-                // Failed to read value
-                Log.w(TAG, "Failed to read value.", error.toException());
-                Utility.generalError(ListAvailableEventsActivity.this, error.getMessage());
-            }
-        });
+        // set the RecyclerView parameters
+        LinearLayoutManager llm = new LinearLayoutManager(ListAvailableEventsActivity.this);
+        llm.setOrientation(LinearLayoutManager.VERTICAL);
+
+        mRecyclerView.setLayoutManager(llm);
+        mRecyclerView.setHasFixedSize(true);
+
+        // Update the list with the results
+        adapter = new AvailableEventsAdapter(ListAvailableEventsActivity.this, events);
+        mRecyclerView.setAdapter(adapter);
     }
 
+    @Override
+    public void OnGetEventsListFailed(String error) {
+        // Failed to read value
+        Log.w(TAG, error);
+        Utility.generalError(ListAvailableEventsActivity.this, error);
+    }
 }
