@@ -20,6 +20,8 @@ import android.widget.TextView;
 import com.filipe.footballmatch.Models.Event;
 import com.filipe.footballmatch.Models.Person;
 import com.filipe.footballmatch.R;
+import com.filipe.footballmatch.Repositories.EventRepository;
+import com.filipe.footballmatch.Repositories.UserRepository;
 import com.filipe.footballmatch.Utilities.MessageDialog;
 import com.filipe.footballmatch.Utilities.Utility;
 import com.google.android.gms.location.places.ui.PlacePicker;
@@ -35,6 +37,9 @@ import org.parceler.Parcels;
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 
+import butterknife.BindView;
+import butterknife.ButterKnife;
+
 import static android.R.attr.switchMinWidth;
 import static android.R.attr.value;
 import static android.view.View.GONE;
@@ -42,44 +47,69 @@ import static com.filipe.footballmatch.R.id.pickerButton;
 import static com.filipe.footballmatch.R.id.section2;
 import static com.filipe.footballmatch.R.id.section3;
 import static com.filipe.footballmatch.Utilities.Utility.call;
-import static com.google.android.gms.analytics.internal.zzy.c;
-import static com.google.android.gms.analytics.internal.zzy.i;
 
 /**
  * Created by Filipe on 21/01/2017.
  */
 
-public class ViewMatchActivity extends AppCompatActivity {
+public class ViewMatchActivity extends AppCompatActivity implements EventRepository.OnGetEventInfo, UserRepository.OnGetUsersList, EventRepository.OnFinished {
 
     private static final int REMOVE_PLAYER_REQUEST_CODE = 1001;
 
-    private TextView tvName;
-    private TextView tvVenueName;
-    private TextView tvAddress;
-    private RelativeLayout lPhone;
-    private TextView tvPhone;
-    private TextView tvNumberOfPlayers;
-    private TextView tvEventDate;
-    private TextView tvEventTime;
-    private LinearLayout lEventPlayers;
-    private LinearLayout lSeparator;
+    @BindView(R.id.event_name)
+    TextView tvName;
 
-    private TextView callButton;
-    private TextView buttonEditEvent;
-    private TextView buttonCancelEvent;
+    @BindView(R.id.venue_name)
+    TextView tvVenueName;
+
+    @BindView(R.id.venue_address)
+    TextView tvAddress;
+
+    @BindView(R.id.section3)
+    RelativeLayout lPhone;
+
+    @BindView(R.id.venue_phone)
+    TextView tvPhone;
+
+    @BindView(R.id.event_number_of_players)
+    TextView tvNumberOfPlayers;
+
+    @BindView(R.id.event_date)
+    TextView tvEventDate;
+
+    @BindView(R.id.event_time)
+    TextView tvEventTime;
+
+    @BindView(R.id.event_players_list)
+    LinearLayout lEventPlayers;
+
+    @BindView(R.id.separator_4)
+    LinearLayout lSeparator;
+
+    @BindView(R.id.callButton)
+    TextView callButton;
+
+    @BindView(R.id.buttonEditEvent)
+    TextView buttonEditEvent;
+
+    @BindView(R.id.buttonCancelEvent)
+    TextView buttonCancelEvent;
 
     FirebaseDatabase database = FirebaseDatabase.getInstance();
     DatabaseReference myRef;
 
     public static final String TAG = ViewMatchActivity.class.getSimpleName();
+    UserRepository userRepository = new UserRepository();
+    EventRepository eventRepository = new EventRepository();
 
     String id;
     String eventId;
     Event event;
     ArrayList<String> playerIdList = new ArrayList<>();
     private ArrayList<Person> playerList = new ArrayList<>();
-    int index;
     boolean isFirstTime = true;
+
+    String message;
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
@@ -87,164 +117,28 @@ public class ViewMatchActivity extends AppCompatActivity {
 
         // The content layout of screen is set
         setContentView(R.layout.activity_view_match);
+        ButterKnife.bind(this);
 
         // The action bar title is customized
         getSupportActionBar().setDisplayOptions(ActionBar.DISPLAY_SHOW_CUSTOM);
         getSupportActionBar().setCustomView(R.layout.action_bar);
 
-        // The layout is now built
-        // First, the TextViews that will show the information to the user
-        tvName = (TextView) findViewById(R.id.event_name);
-        tvVenueName = (TextView) findViewById(R.id.venue_name);
-        tvAddress = (TextView) findViewById(R.id.venue_address);
-        tvNumberOfPlayers = (TextView) findViewById(R.id.event_number_of_players);
-        tvPhone = (TextView) findViewById(R.id.venue_phone);
-        tvEventDate = (TextView) findViewById(R.id.event_date);
-        tvEventTime = (TextView) findViewById(R.id.event_time);
-
-        // Then, the layout that will only appears if the info exists
-        lPhone = (RelativeLayout) findViewById(section3);
-        lSeparator = (LinearLayout) findViewById(R.id.separator_4);
-
-        // The list of players in the match will increase dynamically,
-        // during app execution
-        lEventPlayers = (LinearLayout) findViewById(R.id.event_players_list);
-
-        // Finally, the TextViews that will act as ViewMatchActivity screen buttons are set
-        callButton = (TextView) findViewById(R.id.callButton);
-        buttonEditEvent = (TextView) findViewById(R.id.buttonEditEvent);
-        buttonCancelEvent = (TextView) findViewById(R.id.buttonCancelEvent);
         // An instance of FirebaseDatabase is set
         myRef = database.getReference("Event/");
 
-        // The actual user id is recovered from the device local storage...
-        SharedPreferences saved_values = PreferenceManager.getDefaultSharedPreferences(getApplicationContext());
-        id = saved_values.getString(getString(R.string.user_id_SharedPref), "");
+        // The actual user id is recovered
+        id = userRepository.getUidCurrentUser();
 
         // The eventId is recovered and depends on the item clicked by the user on the ListAvailableEventsActivity
         eventId = getIntent().getStringExtra("eventKey");
 
         // Read from the database
-        myRef.child(eventId).addListenerForSingleValueEvent(new ValueEventListener() {
-
-            @Override
-            public void onDataChange(DataSnapshot dataSnapshot) {
-
-                // This method is called once with the initial value and again
-                // whenever data at this location is updated.
-                event = dataSnapshot.getValue(Event.class);
-                Log.d(TAG, "Value is: " + value);
-
-                // The TextViews are, then, populated
-                tvName.setText(event.getEventName());
-                tvVenueName.setText(event.getName());
-                tvAddress.setText(event.getAddress());
-
-                if (event.getPhone() != null && !event.getPhone().isEmpty()) {
-                    lPhone.setVisibility(View.VISIBLE);
-                    tvPhone.setText(event.getPhone());
-
-                    callButton.setOnClickListener(new View.OnClickListener() {
-                        @Override
-                        public void onClick(View v) {
-                            Utility.call(ViewMatchActivity.this, tvPhone.getText().toString());
-                        }
-                    });
-                }
-
-                tvNumberOfPlayers.setText(event.getNumberOfPlayers());
-
-                SimpleDateFormat sdf = new SimpleDateFormat("dd-MM-yyyy");
-                String formattedDate = sdf.format(event.getDate());
-                tvEventDate.setText(formattedDate);
-
-                sdf = new SimpleDateFormat("hh:mm a");
-                String formattedTime = sdf.format(event.getDate());
-
-                tvEventTime.setText(formattedTime);
-
-                if (event.getPlayersIdList() != null
-                        && !event.getPlayersIdList().isEmpty()) {
-                    lSeparator.setVisibility(View.VISIBLE);
-                    playerIdList = event.getPlayersIdList();
-
-                    if (isFirstTime) {
-                        isFirstTime = false;
-                        for (String id : playerIdList) {
-                            getIdInfo(id);
-                        }
-                    }
-                } else {
-                    lSeparator.setVisibility(GONE);
-                }
-
-                // If the chosen match was created by the actual user of the app, it can be edited
-                // When the button edit match is clicked, the data about the match is wrapped and sent
-                // to the EditMatchActivity
-                if (event.getCreator() != null && event.getCreator().equals(id)) {
-                    changeButton(0);
-                } else {
-                    if (!playerIdList.contains(id)) {
-                        changeButton(1);
-                    } else {
-                        changeButton(2);
-                    }
-
-                }
-
-            }
-
-            @Override
-            public void onCancelled(DatabaseError error) {
-                // Failed to read value
-                Log.w(TAG, "Failed to read value.", error.toException());
-            }
-        });
+        eventRepository.fetchEventInfo(eventId, this);
 
     }
 
-    // For every player added to the event, the app search for its info in the database...
-    public void getIdInfo(final String id) {
-
-        DatabaseReference myRef = database.getReference("Person/");
-
-        // Read from the database
-        myRef.addListenerForSingleValueEvent(new ValueEventListener() {
-
-            @Override
-            public void onDataChange(DataSnapshot dataSnapshot) {
-
-                for (DataSnapshot dsp : dataSnapshot.getChildren()) {
-                    if (dsp.getKey() != null
-                            && dsp.getKey().equals(id)) {
-                        Person temp = dsp.getValue(Person.class);
-                        temp.setUserKey(dsp.getKey());
-                        playerList.add(temp);
-                        populatePlayerLayout();
-                        Log.d(TAG, "Value is: " + value);
-                    } else if (dsp.getValue(Person.class).getOldKey() != null
-                            && dsp.getValue(Person.class).getOldKey().equals(id)) {
-                        Person temp = dsp.getValue(Person.class);
-                        temp.setUserKey(dsp.getKey());
-                        playerList.add(temp);
-                        populatePlayerLayout();
-                        Log.d(TAG, "Value is: " + value);
-                    }
-                }
-            }
-
-            @Override
-            public void onCancelled(DatabaseError error) {
-                // Failed to read value
-                Log.w(TAG, "Failed to read value.", error.toException());
-                Utility.generalError(ViewMatchActivity.this, error.getMessage());
-            }
-
-        });
-
-    }
-
-    // ... and populate the layout with the player name and preferred position
+    // For every player added to the event, the app search for its info in the database
+    // and populate the layout with the player name and preferred position
     public void populatePlayerLayout() {
 
         int i = playerList.size() - 1;
@@ -256,40 +150,25 @@ public class ViewMatchActivity extends AppCompatActivity {
         playerName.setText(playerList.get(i).getName());
         playerPreferredPosition.setText(" (" + playerList.get(i).getPreferredPosition() + ")");
 
-        view.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
+        view.setOnClickListener(v -> {
 
-                Intent intent = new Intent(ViewMatchActivity.this, ViewProfileActivity.class);
-                intent.putExtra("isFromViewMatch", true);
-                intent.putExtra("userKey", v.getTag().toString());
-                ViewMatchActivity.this.startActivityForResult(intent, REMOVE_PLAYER_REQUEST_CODE);
+            Intent intent = new Intent(ViewMatchActivity.this, ViewProfileActivity.class);
+            intent.putExtra("isFromViewMatch", true);
+            intent.putExtra("userKey", v.getTag().toString());
+            ViewMatchActivity.this.startActivityForResult(intent, REMOVE_PLAYER_REQUEST_CODE);
 
-            }
         });
     }
 
     // Update the match players list
     public void updateMatch(String message) {
 
-        //  Write a message to the database
-        FirebaseDatabase database = FirebaseDatabase.getInstance();
-        DatabaseReference myRef = database.getReference("Event");
+        this.message = message;
 
         // Adding values
         event.setPlayersIdList(playerIdList);
 
-        myRef.child(event.getEventKey()).setValue(event);
-
-        final MessageDialog dialog = new MessageDialog(ViewMatchActivity.this, message, R.string.dialog_edit_ok_text, -1, -1);
-        dialog.setCancelable(false);
-        dialog.show();
-        dialog.okButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                dialog.dismiss();
-            }
-        });
+        eventRepository.saveEvent(event, this);
     }
 
     @Override
@@ -326,134 +205,205 @@ public class ViewMatchActivity extends AppCompatActivity {
     }
 
     public void changeButton(int option) {
-        switch(option) {
+        switch (option) {
             // case 0: user is the creator of the event
             case 0:
-                buttonEditEvent.setOnClickListener(new View.OnClickListener() {
-                    @Override
-                    public void onClick(View v) {
-                        if (ActivityCompat.checkSelfPermission(ViewMatchActivity.this, Manifest.permission.ACCESS_FINE_LOCATION) != PackageManager.PERMISSION_GRANTED
-                                || ActivityCompat.checkSelfPermission(ViewMatchActivity.this, Manifest.permission.ACCESS_COARSE_LOCATION) != PackageManager.PERMISSION_GRANTED) {
-                            //Can add more as per requirement  
+                buttonEditEvent.setOnClickListener(v -> {
+                    if (ActivityCompat.checkSelfPermission(ViewMatchActivity.this, Manifest.permission.ACCESS_FINE_LOCATION) != PackageManager.PERMISSION_GRANTED
+                            || ActivityCompat.checkSelfPermission(ViewMatchActivity.this, Manifest.permission.ACCESS_COARSE_LOCATION) != PackageManager.PERMISSION_GRANTED) {
+                        //Can add more as per requirement  
 
-                            ActivityCompat.requestPermissions(ViewMatchActivity.this,
-                                    new String[]{Manifest.permission.ACCESS_FINE_LOCATION, Manifest.permission.ACCESS_COARSE_LOCATION}, 121);
-                        } else {
-                            Intent intent = new Intent(ViewMatchActivity.this, EditMatchActivity.class);
-                            intent.putExtra("event", Parcels.wrap(event));
-                            ViewMatchActivity.this.startActivity(intent);
-                        }
+                        ActivityCompat.requestPermissions(ViewMatchActivity.this,
+                                new String[]{Manifest.permission.ACCESS_FINE_LOCATION, Manifest.permission.ACCESS_COARSE_LOCATION}, 121);
+                    } else {
+                        Intent intent = new Intent(ViewMatchActivity.this, EditMatchActivity.class);
+                        intent.putExtra("event", Parcels.wrap(event));
+                        ViewMatchActivity.this.startActivity(intent);
                     }
                 });
                 buttonCancelEvent.setVisibility(View.VISIBLE);
-                buttonCancelEvent.setOnClickListener(new View.OnClickListener() {
-                    @Override
-                    public void onClick(View v) {
-                        final MessageDialog dialog = new MessageDialog(ViewMatchActivity.this, R.string.cancel_event_confirmation_message,
-                                -1, R.string.dialog_edit_no_text, R.string.dialog_edit_yes_text);
-                        dialog.setCancelable(false);
-                        dialog.show();
-                        dialog.noButton.setOnClickListener(new View.OnClickListener() {
-                            @Override
-                            public void onClick(View view) {
-                                dialog.dismiss();
-                            }
+                buttonCancelEvent.setOnClickListener(v -> {
+                    final MessageDialog dialog = new MessageDialog(ViewMatchActivity.this, R.string.cancel_event_confirmation_message,
+                            -1, R.string.dialog_edit_no_text, R.string.dialog_edit_yes_text);
+                    dialog.setCancelable(false);
+                    dialog.show();
+                    dialog.noButton.setOnClickListener(view -> dialog.dismiss());
+                    dialog.yesButton.setOnClickListener(view -> {
+
+                        // remove the current event from the database
+                        myRef.child(eventId).removeValue();
+
+                        dialog.dismiss();
+
+                        final MessageDialog dialog1 = new MessageDialog(ViewMatchActivity.this, R.string.success_cancel_match, R.string.dialog_edit_ok_text, -1, -1);
+                        dialog1.setCancelable(false);
+                        dialog1.show();
+                        dialog1.okButton.setOnClickListener(view1 -> {
+                            Intent intent = new Intent(ViewMatchActivity.this, MainMenuActivity.class);
+                            ViewMatchActivity.this.startActivity(intent);
+
+                            dialog1.cancel();
                         });
-                        dialog.yesButton.setOnClickListener(new View.OnClickListener() {
-                            @Override
-                            public void onClick(View view) {
-
-                                // remove the current event from the database
-                                myRef.child(eventId).removeValue();
-
-                                dialog.dismiss();
-
-                                final MessageDialog dialog = new MessageDialog(ViewMatchActivity.this, R.string.success_cancel_match, R.string.dialog_edit_ok_text, -1, -1);
-                                dialog.setCancelable(false);
-                                dialog.show();
-                                dialog.okButton.setOnClickListener(new View.OnClickListener() {
-                                    @Override
-                                    public void onClick(View view) {
-                                        Intent intent = new Intent(ViewMatchActivity.this, MainMenuActivity.class);
-                                        ViewMatchActivity.this.startActivity(intent);
-
-                                        dialog.cancel();
-                                    }
-                                });
-                            }
-                        });
-                    }
+                    });
                 });
                 break;
             // case 1: player is not registered in the event
             case 1:
                 buttonEditEvent.setText(getString(R.string.join_match_button));
-                buttonEditEvent.setOnClickListener(new View.OnClickListener() {
-                    @Override
-                    public void onClick(View v) {
+                buttonEditEvent.setOnClickListener(v -> {
 
-                        int maxPlayers = 0;
-                        switch (event.getNumberOfPlayers()) {
-                            case "10 (5x2)":
-                                maxPlayers = 10;
-                                break;
-                            case "14 (7x2)":
-                                maxPlayers = 14;
-                                break;
-                            case "22 (11x2)":
-                                maxPlayers = 22;
-                                break;
-                        }
-
-                        if (playerIdList.size() >= maxPlayers){
-                            Utility.generalError(ViewMatchActivity.this, getString(R.string.error_join_match));
-                        } else {
-                            // add the current user to the match players list
-                            playerIdList.add(id);
-                            getIdInfo(id);
-
-                            //update the database
-                            updateMatch(getString(R.string.success_join_match));
-                            changeButton(2);
-
-
-                        }
-
-
+                    int maxPlayers = 0;
+                    switch (event.getNumberOfPlayers()) {
+                        case "10 (5x2)":
+                            maxPlayers = 10;
+                            break;
+                        case "14 (7x2)":
+                            maxPlayers = 14;
+                            break;
+                        case "22 (11x2)":
+                            maxPlayers = 22;
+                            break;
                     }
+
+                    if (playerIdList.size() >= maxPlayers) {
+                        Utility.generalError(ViewMatchActivity.this, getString(R.string.error_join_match));
+                    } else {
+                        // add the current user to the match players list
+                        playerIdList.add(id);
+                        userRepository.fetchListUsers(this);
+
+                        //update the database
+                        updateMatch(getString(R.string.success_join_match));
+                        changeButton(2);
+                    }
+
                 });
                 break;
             // case 2: player is already registered in the event
             case 2:
                 buttonEditEvent.setText(getString(R.string.leave_event_button));
-                buttonEditEvent.setOnClickListener(new View.OnClickListener() {
-                    @Override
-                    public void onClick(View v) {
-                        final MessageDialog dialog = new MessageDialog(ViewMatchActivity.this, R.string.leave_match_confirmation_message,
-                                -1, R.string.dialog_edit_no_text, R.string.dialog_edit_yes_text);
-                        dialog.setCancelable(false);
-                        dialog.show();
-                        dialog.noButton.setOnClickListener(new View.OnClickListener() {
-                            @Override
-                            public void onClick(View view) {
-                                dialog.dismiss();
-                            }
-                        });
-                        dialog.yesButton.setOnClickListener(new View.OnClickListener() {
-                            @Override
-                            public void onClick(View view) {
+                buttonEditEvent.setOnClickListener(v -> {
+                    final MessageDialog dialog = new MessageDialog(ViewMatchActivity.this, R.string.leave_match_confirmation_message,
+                            -1, R.string.dialog_edit_no_text, R.string.dialog_edit_yes_text);
+                    dialog.setCancelable(false);
+                    dialog.show();
+                    dialog.noButton.setOnClickListener(view -> dialog.dismiss());
+                    dialog.yesButton.setOnClickListener(view -> {
 
-                                // remove the current user from the match players id list
-                                removePlayer(id);
-                                dialog.dismiss();
-                                changeButton(1);
+                        // remove the current user from the match players id list
+                        removePlayer(id);
+                        dialog.dismiss();
+                        changeButton(1);
 
-                            }
-                        });
-                    }
+                    });
                 });
                 break;
         }
     }
 
+    @Override
+    public void OnGetEventInfoSuccess(Event match) {
+
+        event = match;
+
+        // The TextViews are, then, populated
+        tvName.setText(event.getEventName());
+        tvVenueName.setText(event.getName());
+        tvAddress.setText(event.getAddress());
+
+        if (event.getPhone() != null && !event.getPhone().isEmpty()) {
+            lPhone.setVisibility(View.VISIBLE);
+            tvPhone.setText(event.getPhone());
+
+            callButton.setOnClickListener(v -> Utility.call(ViewMatchActivity.this, tvPhone.getText().toString()));
+        }
+
+        tvNumberOfPlayers.setText(event.getNumberOfPlayers());
+
+        SimpleDateFormat sdf = new SimpleDateFormat("dd-MM-yyyy");
+        String formattedDate = sdf.format(event.getDate());
+        tvEventDate.setText(formattedDate);
+
+        sdf = new SimpleDateFormat("hh:mm a");
+        String formattedTime = sdf.format(event.getDate());
+
+        tvEventTime.setText(formattedTime);
+
+        if (event.getPlayersIdList() != null
+                && !event.getPlayersIdList().isEmpty()) {
+            lSeparator.setVisibility(View.VISIBLE);
+            playerIdList = event.getPlayersIdList();
+
+            if (isFirstTime) {
+                isFirstTime = false;
+                userRepository.fetchListUsers(this);
+            }
+        } else {
+            lSeparator.setVisibility(GONE);
+        }
+
+        // If the chosen match was created by the actual user of the app, it can be edited
+        // When the button edit match is clicked, the data about the match is wrapped and sent
+        // to the EditMatchActivity
+        if (event.getCreator() != null && event.getCreator().equals(id)) {
+            changeButton(0);
+        } else {
+            if (!playerIdList.contains(id)) {
+                changeButton(1);
+            } else {
+                changeButton(2);
+            }
+
+        }
+    }
+
+    @Override
+    public void OnGetEventInfoFailed(String error) {
+        // Failed to read value
+        Log.e(TAG, error);
+        Utility.generalError(ViewMatchActivity.this, error);
+    }
+
+    @Override
+    public void OnGetUsersListSuccess(DataSnapshot dataSnapshot) {
+        for (DataSnapshot dsp : dataSnapshot.getChildren()) {
+            if (dsp.getKey() != null
+                    && dsp.getKey().equals(id)) {
+                Person temp = dsp.getValue(Person.class);
+                temp.setUserKey(dsp.getKey());
+                playerList.add(temp);
+                populatePlayerLayout();
+                Log.d(TAG, "Value is: " + value);
+            } else if (dsp.getValue(Person.class).getOldKey() != null
+                    && dsp.getValue(Person.class).getOldKey().equals(id)) {
+                Person temp = dsp.getValue(Person.class);
+                temp.setUserKey(dsp.getKey());
+                playerList.add(temp);
+                populatePlayerLayout();
+                Log.d(TAG, "Value is: " + value);
+            }
+        }
+    }
+
+    @Override
+    public void OnGetUsersListFailed(String error) {
+        // Failed to read value
+        Log.e(TAG, error);
+        Utility.generalError(ViewMatchActivity.this, error);
+    }
+
+    @Override
+    public void onEventSaveSuccess() {
+        final MessageDialog dialog = new MessageDialog(ViewMatchActivity.this, message, R.string.dialog_edit_ok_text, -1, -1);
+        dialog.setCancelable(false);
+        dialog.show();
+        dialog.okButton.setOnClickListener(view -> dialog.dismiss());
+    }
+
+    @Override
+    public void onEventSaveFailed(String exception) {
+        // Failed to read value
+        Log.e(TAG, exception);
+        Utility.generalError(ViewMatchActivity.this, exception);
+    }
 }
