@@ -17,6 +17,7 @@ import android.widget.TextView;
 
 import com.filipe.footballmatch.Models.Person;
 import com.filipe.footballmatch.R;
+import com.filipe.footballmatch.Repositories.UserRepository;
 import com.filipe.footballmatch.Utilities.Utility;
 import com.google.android.gms.tasks.OnFailureListener;
 import com.google.android.gms.tasks.OnSuccessListener;
@@ -34,30 +35,45 @@ import org.parceler.Parcels;
 import java.io.File;
 import java.io.IOException;
 
+import butterknife.BindView;
+import butterknife.ButterKnife;
+
 import static android.R.attr.value;
 
 /**
  * Created by Filipe on 21/01/2017.
  */
 
-public class ViewProfileActivity extends AppCompatActivity {
+public class ViewProfileActivity extends AppCompatActivity implements UserRepository.OnGetUserImage, UserRepository.OnGetUserInfo{
 
-    private ImageView profilePicture;
-    private TextView tvName;
-    private TextView tvAge;
-    private TextView tvPreferredPosition;
-    private TextView tvContactNumber;
-    private TextView tvEmail;
+    @BindView(R.id.profile_picture_iv)
+    ImageView profilePicture;
 
+    @BindView(R.id.user_name)
+    TextView tvName;
+
+    @BindView(R.id.user_age)
+    TextView tvAge;
+
+    @BindView(R.id.user_preferred_position)
+    TextView tvPreferredPosition;
+
+    @BindView(R.id.user_contact_number)
+    TextView tvContactNumber;
+
+    @BindView(R.id.user_email)
+    TextView tvEmail;
+
+    @BindView(R.id.callButton)
     TextView buttonCallPlayer;
+
+    @BindView(R.id.buttonEditProfile)
     TextView buttonEditProfile;
 
-    StorageReference profileImageRef;
-
     public static final String TAG = ViewProfileActivity.class.getSimpleName();
+    UserRepository userRepository = new UserRepository();
 
     String id;
-    File localFile = null;
     Person person;
     Boolean isFromViewMatch;
 
@@ -67,6 +83,7 @@ public class ViewProfileActivity extends AppCompatActivity {
 
         // The content layout of screen is set
         setContentView(R.layout.activity_view_profile);
+        ButterKnife.bind(this);
 
         // The action bar title is customized
         getSupportActionBar().setDisplayOptions(ActionBar.DISPLAY_SHOW_CUSTOM);
@@ -76,22 +93,8 @@ public class ViewProfileActivity extends AppCompatActivity {
         // (If it is from view match or directly from the main menu)
         isFromViewMatch = getIntent().getBooleanExtra("isFromViewMatch", false);
 
-        // The layout is now built
-        // First, the TextViews that will show the information to the user
-        profilePicture = (ImageView) findViewById(R.id.profile_picture_iv);
-        tvName = (TextView) findViewById(R.id.user_name);
-        tvAge = (TextView) findViewById(R.id.user_age);
-        tvPreferredPosition = (TextView) findViewById(R.id.user_preferred_position);
-        tvContactNumber = (TextView) findViewById(R.id.user_contact_number);
-        tvEmail = (TextView) findViewById(R.id.user_email);
-
-        // Then, the TextView that will act as ViewProfileActivity screen button is set
-        buttonEditProfile = (TextView) findViewById(R.id.buttonEditProfile);
-        buttonCallPlayer = (TextView) findViewById(R.id.callButton);
-
-        // The actual user id is recovered from the device local storage...
-        SharedPreferences saved_values = PreferenceManager.getDefaultSharedPreferences(getApplicationContext());
-        id = saved_values.getString(getString(R.string.user_id_SharedPref), "");
+        // The actual user id is recovered
+        id = userRepository.getUidCurrentUser();
 
         // ... and compared with the id that is coming with the intent. If they are equal or if the intent
         //  is not bringing any info about the userKey, it means that the profile that are being shown
@@ -104,129 +107,75 @@ public class ViewProfileActivity extends AppCompatActivity {
             // If the chosen profile is the same as of the actual user of the app, it can be edited
             // When the button edit profile is clicked, the data about the user is wrapped and sent
             // to the EditProfileActivity
-            buttonEditProfile.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View v) {
-                    Intent intent = new Intent(ViewProfileActivity.this, EditProfileActivity.class);
-                    intent.putExtra("person", Parcels.wrap(person));
-                    ViewProfileActivity.this.startActivity(intent);
-                    finish();
-                }
+            buttonEditProfile.setOnClickListener(v -> {
+                Intent intent = new Intent(ViewProfileActivity.this, EditProfileActivity.class);
+                intent.putExtra("person", Parcels.wrap(person));
+                ViewProfileActivity.this.startActivity(intent);
+                finish();
             });
-        }
-        else {
+        } else {
             id = getIntent().getStringExtra("userKey");
         }
 
-        // Now, the app recovers the player image from the cloud storage
-
-        // Get instance of Storage
-        FirebaseStorage storage = FirebaseStorage.getInstance();
-
-        // Create a storage reference from our app
-        StorageReference storageRef = storage.getReferenceFromUrl("gs://football-match-a60fb.appspot.com");
-
+        // Now, the app recovers the player image from the cloud storage and the player info from the database
         if (!id.isEmpty()) {
-            // Create a reference to 'images/id.jpg'
-            profileImageRef = storageRef.child("images/" + id + ".jpg");
+            userRepository.fetchUserImage(id, this);
+            userRepository.fetchUserInfo(id, this);
+        }
+    }
 
-            // Try create a file with the recovered picture
-            try {
-                localFile = File.createTempFile(id, "jpg");
-            } catch (IOException e) {
-                e.printStackTrace();
+    public void populateFields() {
+        // The TextViews are, then, populated
+        tvName.setText(person.getName());
+        tvAge.setText(Integer.toString(person.getAge()));
+        tvPreferredPosition.setText(person.getPreferredPosition());
+
+        // If the contact number exists, then the field is populated...
+        if (person.getContactNumber() != null
+                && !person.getContactNumber().isEmpty()) {
+            tvContactNumber.setText(person.getContactNumber());
+
+            // and, if the user is not seeing its own profile, a button call appears,
+            // allowing the player to be contacted
+            if (buttonEditProfile.getVisibility() == View.GONE) {
+                buttonCallPlayer.setVisibility(View.VISIBLE);
+                buttonCallPlayer.setOnClickListener(v -> Utility.call(ViewProfileActivity.this, person.getContactNumber()));
             }
         }
 
-        if (localFile != null) {
-            profileImageRef.getFile(localFile).addOnSuccessListener(new OnSuccessListener<FileDownloadTask.TaskSnapshot>() {
-                @Override
-                public void onSuccess(FileDownloadTask.TaskSnapshot taskSnapshot) {
-                    Bitmap myImg = BitmapFactory.decodeFile(localFile.getPath());
-                    profilePicture.setImageBitmap(myImg);
-                }
-            }).addOnFailureListener(new OnFailureListener() {
-                @Override
-                public void onFailure(@NonNull Exception exception) {
-                    // Handle any errors
-                }
+        tvEmail.setText(person.getEmail());
+
+        // If the user reach this screen from the view match flow, a new option appears
+        // this option allows the user to remove the player from the match
+        if (isFromViewMatch && buttonEditProfile.getVisibility() == View.GONE) {
+            buttonEditProfile.setText(getString(R.string.remove_player_button));
+            buttonEditProfile.setVisibility(View.VISIBLE);
+            buttonEditProfile.setOnClickListener(v -> {
+                Intent returnIntent = new Intent(ViewProfileActivity.this, ViewMatchActivity.class);
+                returnIntent.putExtra("playerId", person.getUserKey());
+                setResult(Activity.RESULT_OK, returnIntent);
+                finish();
             });
         }
-
-
-
-        if (!id.isEmpty()) {
-            // An instance of FirebaseDatabase is set
-            FirebaseDatabase database = FirebaseDatabase.getInstance();
-            DatabaseReference myRef = database.getReference("Person/");
-
-            // Read from the database
-            myRef.child(id).addListenerForSingleValueEvent(new ValueEventListener() {
-
-                @Override
-                public void onDataChange(final DataSnapshot dataSnapshot) {
-                    // This method is called once with the initial value and again
-                    // whenever data at this location is updated.
-                    person = dataSnapshot.getValue(Person.class);
-                    Log.d(TAG, "Value is: " + value);
-
-                    // The TextViews are, then, populated
-                    tvName.setText(person.getName());
-                    tvAge.setText(Integer.toString(person.getAge()));
-                    tvPreferredPosition.setText(person.getPreferredPosition());
-
-                    // If the contact number exists, then the field is populated...
-                    if (person.getContactNumber() != null
-                        && !person.getContactNumber().isEmpty()) {
-                        tvContactNumber.setText(person.getContactNumber());
-
-                        // and, if the user is not seeing its own profile, a button call appears,
-                        // allowing the player to be contacted
-                        if (buttonEditProfile.getVisibility() == View.GONE) {
-                            buttonCallPlayer.setVisibility(View.VISIBLE);
-                            buttonCallPlayer.setOnClickListener(new View.OnClickListener() {
-                                @Override
-                                public void onClick(View v) {
-                                    Utility.call(ViewProfileActivity.this, person.getContactNumber());
-                                }
-                            });
-                        }
-                    }
-
-                    tvEmail.setText(person.getEmail());
-
-                    // If the user reach this screen from the view match flow, a new option appears
-                    // this option allows the user to remove the player from the match
-                    if (isFromViewMatch && buttonEditProfile.getVisibility() == View.GONE) {
-                        buttonEditProfile.setText(getString(R.string.remove_player_button));
-                        buttonEditProfile.setVisibility(View.VISIBLE);
-                        buttonEditProfile.setOnClickListener(new View.OnClickListener() {
-                            @Override
-                            public void onClick(View v) {
-                                Intent returnIntent = new Intent(ViewProfileActivity.this, ViewMatchActivity.class);
-                                returnIntent.putExtra("playerId",dataSnapshot.getKey());
-                                setResult(Activity.RESULT_OK,returnIntent);
-                                finish();
-                            }
-                        });
-                    }
-
-                }
-
-
-
-                @Override
-                public void onCancelled(DatabaseError error) {
-                    // Failed to read value
-                    Log.w(TAG, "Failed to read value.", error.toException());
-                }
-            });
-
-
-
-        }
-
     }
 
 
+    @Override
+    public void OnGetUserImageSuccess(Bitmap myImg) {
+        profilePicture.setImageBitmap(myImg);
+    }
+
+    @Override
+    public void OnGetUserInfoSuccess(Person user) {
+        person = user;
+
+        populateFields();
+    }
+
+    @Override
+    public void OnGetUserInfoFailed(String error) {
+        // Failed to read value
+        Log.e(TAG, error);
+        Utility.generalError(ViewProfileActivity.this, error);
+    }
 }
